made one consistent way of checking if user is prof, split away a separate lecture section view for students
